+++ src/dashboard/apigateway/apigateway/apps/gateway/admin.py
@@ -1,0 +1,30 @@
+# -*- coding: utf-8 -*-
+#
+# TencentBlueKing is pleased to support the open source community by making
+# 蓝鲸智云 - API 网关(BlueKing - APIGateway) available.
+# Copyright (C) 2017 THL A29 Limited, a Tencent company. All rights reserved.
+# Licensed under the MIT License (the "License"); you may not use this file except
+# in compliance with the License. You may obtain a copy of the License at
+#
+#     http://opensource.org/licenses/MIT
+#
+# Unless required by applicable law or agreed to in writing, software distributed under
+# the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+# either express or implied. See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# We undertake not to change the open source license (MIT license) applicable
+# to the current version of the project delivered to anyone in the future.
+#
+from django.contrib import admin
+
+from .models import GatewayAppBinding
+
+
+class GatewayAppBindingAdmin(admin.ModelAdmin):
+    list_display = ["id", "gateway", "bk_app_code", "updated_time"]
+    search_fields = ["gateway__id", "bk_app_code"]
+    list_filter = ["gateway"]
+
+
+admin.site.register(GatewayAppBinding, GatewayAppBindingAdmin)
