+++ src/dashboard/apigateway/apigateway/core/admin.py
@@ -23,7 +23,6 @@ from apigateway.core.models import (
     BackendService,
     Context,
     Gateway,
-    GatewayAppBinding,
     MicroGateway,
     MicroGatewayReleaseHistory,
     Proxy,
@@ -168,12 +167,6 @@ class SslCertificateBindingAdmin(admin.ModelAdmin):
     list_filter = ["api"]
 
 
-class GatewayAppBindingAdmin(admin.ModelAdmin):
-    list_display = ["id", "gateway", "bk_app_code", "updated_time"]
-    search_fields = ["gateway__id", "bk_app_code"]
-    list_filter = ["gateway"]
-
-
 admin.site.register(Gateway, GatewayAdmin)
 admin.site.register(Stage, StageAdmin)
 admin.site.register(StageItem, StageItemAdmin)
@@ -194,4 +187,3 @@ admin.site.register(MicroGatewayReleaseHistory, MicroGatewayReleaseHistoryAdmin)
 admin.site.register(BackendService, BackendServiceAdmin)
 admin.site.register(SslCertificate, SslCertificateAdmin)
 admin.site.register(SslCertificateBinding, SslCertificateBindingAdmin)
-admin.site.register(GatewayAppBinding, GatewayAppBindingAdmin)
